Modification de int en float pour fonction

+++ fastfood/Menu.cs
@@ -63,7 +63,7 @@ namespace fastfood
         /// <returns></returns>
         public double CalculerPrixMenu(Article[] tableauArticle)
         {
-            int somme = 0;
+            float somme = 0;
             for(int i = 0; i < tableauArticle.Length; i++)
             {
                 somme = somme + tableauArticle[i].Prix;
